+++ History.md
@@ -1,7 +1,12 @@
 # History
 
 
-## Up to 0.0.63
+## 0.0.64 - 2018-12-11
+
+Auto launch the wallet in the default browser at start  
+(unless debug mode)
+
+## Up to 0.0.63 - 2018-12-10 and before
 
 Proof of concept, unreleased versions.
 0.0.63 Is the first Beta candidate with minimal features